Modify eventlisteners int onclik
